split commands.js into multiple files
