--- package.json
@@ -1,26 +1,0 @@
-{
-    "name": "pesu",
-    "version": "1.0.0",
-    "description": "PESU Discord Bot rewrite in discord.js",
-    "main": "index.js",
-    "dependencies": {
-        "discord.js-commando": "^0.12.3",
-        "discord.js": "^13.3.1",
-        "dotenv": "^10.0.0",
-        "eslint": "^8.5.0"
-    },
-    "devDependencies": {},
-    "scripts": {
-        "test": "cd src && node index.js"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git@github.com:sach-12/pesu-bot-js.git"
-    },
-    "keywords": [],
-    "author": "",
-    "bugs": {
-        "url": "https://github.com/sach-12/pesu-bot-js/issues"
-    },
-    "homepage": "https://github.com/sach-12/pesu-bot-js#readme"
-}
