+++ src/client-commands/misc.js
@@ -1,0 +1,16 @@
+// commands anyone can use
+const config = require('../config.json');
+const clientInfo = require("./clientHelper");
+
+class Misc {
+    constructor() {
+        this.commands = [
+
+        ];
+    }
+
+    
+}
+const misc = new Misc()
+
+module.exports = misc
